model/requirements.txt
torch
ultralytics
fastapi
uvicorn
pillow
aiofiles
python-multipart